WIP: unpickler: reconstruct tuples to resolve cyclical references
Tuples are read-only, so we are often in a situation where
the final object where it will be stored has not yet been
constructed, and thus the tuple ends up with references to
proxies.

Add a final pass that reconstructs tuples with proxies and
restitches the tuples into their containing objects.
The unpickler keeps track of when it seems tuples of interest
so that it can resolve them in the final phase before returning
from restore().

Related-to: #248